Adjust editor parents to support flush embeds.
